Transcribe Video Function/requirements.txt
functions-framework
google-cloud-videointelligence
flask